requirements.txt
asgiref==3.7.2
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
Markdown==3.5
Pillow==10.1.0
psycopg2==2.9.9
pytz==2023.3.post1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
tzdata==2023.3
